Support only Bech32M.
Bitcoin Developers realized that there was an issue with the original Bech32 Implementation, and fixed it with the Bech32M revision. This implies calculating Checksums in another way.
In Riecoin, Bech32M is retroactively applied to Version 0 Addresses as well (starting with 24.12), simplifying implementations as we only have one way to calculate the Checksum.

Signed-off-by: Pttn <28868425+Pttn@users.noreply.github.com>